[#40] Refactor : PriceSlider용 데이터 생성함수를 util로 분리 & 호출위치 변경
- PriceSlider를 그리기 위한 데이터를 생성하는 함수 (generateFormattedPrices)를 util로 분리
- components/Filter/Price/Price.jsx 에서 호출하여 사용하도록 변경
