+++ FE/src/components/Filter/Price/PriceModal.jsx
@@ -3,11 +3,11 @@ import Text from "Styles/Text";
 import PriceSlider from "Components/PriceSlider/PriceSlider";
 import Modal from "../Modal";
 
-const PriceModal = ({ setToggle, isDateSelected }) => {
+const PriceModal = ({ setToggle, isDateSelected, data = null }) => {
   const NO_DATE_MESSAGE = "체크인/체크아웃 날짜를 선택하면\n 요금을 확인할 수 있습니다.";
 
   const modalContent = isDateSelected ? (
-    <PriceSlider />
+    <PriceSlider data={data} />
   ) : (
     NO_DATE_MESSAGE.split("\n").map(chunk => (
       <Text key={chunk} fontSize="md">
@@ -18,7 +18,6 @@ const PriceModal = ({ setToggle, isDateSelected }) => {
 
   const modalOption = {
     contents: modalContent,
-    hasContents: null,
     clearHandler: null,
     toggleHandler: setToggle,
     withButtons: isDateSelected,
